+++ src/view.c
@@ -25,7 +25,7 @@ void view_update_cursor_y(View *v)
 void view_update_cursor_x(View *v)
 {
     unsigned int tw = v->buffer->options.tab_width;
-    long idx = 0;
+    size_t idx = 0;
     LineRef lr;
     int c = 0;
     int w = 0;
@@ -142,7 +142,7 @@ bool view_can_close(View *v)
 char *view_get_word_under_cursor(View *v)
 {
     LineRef lr;
-    long i, ei, si = fetch_this_line(&v->cursor, &lr);
+    size_t i, ei, si = fetch_this_line(&v->cursor, &lr);
 
     while (si < lr.size) {
         i = si;
